add second table

+++ db/migrate/20181105231102_ilearn_tax_information.rb
@@ -4,14 +4,13 @@ class IlearnTaxInformation < ActiveRecord::Migration[5.2]
       t.bigint :district_number
       t.string :district_name
       t.bigint :real_eav
-      t.integer :real_per_pupil
+      t.bigint :real_per_pupil
       t.integer :statewide_eavpp_rank
       t.string :formula_type
       t.integer :total_tax_rate
       t.integer :statewide_ttr_rank
       t.integer :operating_tax_rate
-      t.integer :statewide_otr_rank
-
+      t.bigint :statewide_otr_rank
     end
   end
 end
