+++ db/migrate/20181106123059_create_receipts_revenues.rb
@@ -3,9 +3,9 @@ class CreateReceiptsRevenues < ActiveRecord::Migration[5.2]
     create_table :receipts_revenues do |t|
       t.bigint :district_number
       t.string :district_name
-      t.integer :state_amount
-      t.integer :local_amount
-      t.integer :federal_amount
+      t.bigint :state_amount
+      t.bigint :local_amount
+      t.bigint :federal_amount
       t.bigint :total_amount
       t.string :data_source_url
       t.string :scrape_dev
